requirements.txt
imutils

flask

picamera[array]

Flask-BasicAuth==0.2.0
Rpi.GPIO
Smtp
